requirements.txt
Adafruit-Blinka==8.41.0
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.0
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-requests==4.0.0
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.69.0
Adafruit-PureIO==1.1.11
colorzero==2.0
dht11==0.1.0
gpiozero==2.0
guizero==1.4.0
lgpio==0.2.2.0
pigpio==1.78
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
RPLCD==1.3.1
smbus2==0.4.3
sysv-ipc==1.1.0
typing_extensions==4.12.2
websockets==12.0
